moved window events out of index
